+++ src/main/php/xp/compiler/types/TypeDeclaration.class.php
@@ -127,7 +127,11 @@ class TypeDeclaration extends Types {
         $c= new Constructor();
         $c->modifiers= $member->modifiers;
         foreach ($member->parameters as $p) {
-          $c->parameters[]= $p['type'];
+          $c->parameters[]= new Parameter(
+            $p['name'],
+            $p['type'],
+            isset($p['default']) ? $p['default'] : null
+          );
         }
         $c->holder= $this;
         return $c;
@@ -163,7 +167,11 @@ class TypeDeclaration extends Types {
         $m->returns= $member->returns;
         $m->modifiers= $member->modifiers;
         foreach ((array)$member->parameters as $p) {
-          $m->parameters[]= $p['type'];
+          $m->parameters[]= new Parameter(
+            $p['name'],
+            $p['type'],
+            isset($p['default']) ? $p['default'] : null
+          );
         }
         $m->holder= $this;
         return $m;
@@ -188,7 +196,11 @@ class TypeDeclaration extends Types {
         $m->returns= $member->returns;
         $m->modifiers= $member->modifiers;
         foreach ((array)$member->parameters as $p) {
-          $m->parameters[]= $p['type'];
+          $m->parameters[]= new Parameter(
+            $p['name'],
+            $p['type'],
+            isset($p['default']) ? $p['default'] : null
+          );
         }
         $m->holder= $this;
 
@@ -225,7 +237,11 @@ class TypeDeclaration extends Types {
         $m->returns= $member->returns;
         $m->modifiers= $member->modifiers;
         foreach ($member->parameters as $p) {
-          $m->parameters[]= $p->type;
+          $m->parameters[]= new Parameter(
+            $p['name'],
+            $p['type'],
+            isset($p['default']) ? $p['default'] : null
+          );
         }
         $m->holder= $this;
         return $m;