+++ ChangeLog.md
@@ -3,6 +3,9 @@ XP Compiler ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Heads up: Changed xp.compiler.types.Types implementations to return an 
+  array of Parameter objects holding name, type and optional default value 
+  instead of just the parameters' types - (@thekid)
 * Made error message more verbose when syntax cannot be determined from
   file referenced by command line argument - (@thekid)
 * Fixed Fatal error: Class 'xp\compiler\ClassLoader' not found when using